refactor: login flagging process
