+++ pkg/cmd/login/login_unix.go
@@ -1,0 +1,47 @@
+//go:build freebsd || linux
+
+/*
+   Copyright The containerd Authors.
+
+   Licensed under the Apache License, Version 2.0 (the "License");
+   you may not use this file except in compliance with the License.
+   You may obtain a copy of the License at
+
+       http://www.apache.org/licenses/LICENSE-2.0
+
+   Unless required by applicable law or agreed to in writing, software
+   distributed under the License is distributed on an "AS IS" BASIS,
+   WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+   See the License for the specific language governing permissions and
+   limitations under the License.
+*/
+
+package login
+
+import (
+	"fmt"
+	"os"
+	"syscall"
+
+	"golang.org/x/term"
+)
+
+func readPassword() (string, error) {
+	var fd int
+	if term.IsTerminal(syscall.Stdin) {
+		fd = syscall.Stdin
+	} else {
+		tty, err := os.Open("/dev/tty")
+		if err != nil {
+			return "", fmt.Errorf("error allocating terminal: %w", err)
+		}
+		defer tty.Close()
+		fd = int(tty.Fd())
+	}
+	bytePassword, err := term.ReadPassword(fd)
+	if err != nil {
+		return "", fmt.Errorf("error reading password: %w", err)
+	}
+
+	return string(bytePassword), nil
+}
