+++ pkg/cmd/login/login_windows.go
@@ -14,7 +14,7 @@
    limitations under the License.
 */
 
-package main
+package login
 
 import (
 	"fmt"
